Updates to account routes and pages
