+++ app/__init__.py
@@ -8,7 +8,6 @@ from .models import db, User, Account, Transaction
 from .api.user_routes import user_routes
 from .api.auth_routes import auth_routes
 from .api.account_routes import account_routes
-from .api.transaction_routes import transaction_routes
 from .seeds import seed_commands
 from .config import Config
 
@@ -31,7 +30,6 @@ app.config.from_object(Config)
 app.register_blueprint(user_routes, url_prefix='/api/users')
 app.register_blueprint(auth_routes, url_prefix='/api/auth')
 app.register_blueprint(account_routes, url_prefix='/api/accounts')
-app.register_blueprint(transaction_routes, url_prefix='/api/transactions')
 db.init_app(app)
 Migrate(app, db)
