+++ react-app/src/components/AccountFormModal/index.js
@@ -1,0 +1,47 @@
+import React, { useState } from "react";
+import { createAccount } from "../../store/accounts";
+import { useDispatch } from "react-redux";
+import { useModal } from "../../context/Modal";
+import "./AccountForm.css";
+
+function AccountLoginFormModal() {
+  const dispatch = useDispatch();
+  const [ accountName, setAccountName ] = useState("");
+  const [ errors, setErrors ] = useState([]);
+  const { closeModal } = useModal();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    const data = await dispatch(createAccount(accountName));
+    if (data) {
+      setErrors(data);
+    } else {
+        closeModal()
+    }
+  };
+
+  return (
+    <>
+      <h1>Create New Account</h1>
+      <form onSubmit={handleSubmit}>
+        <ul>
+          {errors.map((error, idx) => (
+            <li key={idx}>{error}</li>
+          ))}
+        </ul>
+        <label>
+          Account Name
+          <input
+            type="text"
+            value={accountName}
+            onChange={(e) => setAccountName(e.target.value)}
+            required
+          />
+        </label>
+        <button type="submit">Create</button>
+      </form>
+    </>
+  );
+}
+
+export default AccountLoginFormModal;
